✨ generic generative token filters & added filters to user { generativeTokens}
