+++ src/Utils/Context.ts
@@ -2,7 +2,7 @@ import { createGentkTokPrimarySplitsLoader, createGentkTokSecondarySplitsLoader,
 import { createMarketStatsGenTokLoader } from '../DataLoaders/MarketStats'
 import { createModerationReasonsLoader } from '../DataLoaders/ModerationReason'
 import { createObjktActionsLoader, createObjktActiveListingsLoader, createObjktListingsLoader, createObjktOffersLoader, createObjktRoyaltiesSplitsLoader, createObjktsLoader } from '../DataLoaders/Objkt'
-import { createUsersLoader, createUsersObjktLoader, createUsersCollabContractsLoader, createCollabCollaboratorsLoader, createUsersOffersLoader } from '../DataLoaders/User'
+import { createUsersLoader, createUsersObjktLoader, createUsersCollabContractsLoader, createCollabCollaboratorsLoader, createUsersOffersLoader, createUsersGenerativeTokensLoader } from '../DataLoaders/User'
 import { RequestContext } from '../types/RequestContext'
 
 export const createContext = (req: any, res: any): RequestContext => {
@@ -16,6 +16,7 @@ export const createContext = (req: any, res: any): RequestContext => {
 		userOffersLoader: createUsersOffersLoader(),
 		userCollabContractsLoader: createUsersCollabContractsLoader(),
 		collabCollaboratorsLoader: createCollabCollaboratorsLoader(),
+		usersGenToksLoader: createUsersGenerativeTokensLoader(),
 
     // GENERATIVE TOKEN loaders
     genTokLoader: createGenTokLoader(),
