+++ src/types/RequestContext.ts
@@ -3,7 +3,7 @@ import { createGentkTokPrimarySplitsLoader, createGentkTokSecondarySplitsLoader,
 import { createMarketStatsGenTokLoader } from "../DataLoaders/MarketStats"
 import { createModerationReasonsLoader } from "../DataLoaders/ModerationReason"
 import { createObjktActionsLoader, createObjktActiveListingsLoader, createObjktListingsLoader, createObjktOffersLoader, createObjktRoyaltiesSplitsLoader, createObjktsLoader } from "../DataLoaders/Objkt"
-import { createUsersObjktLoader, createUsersLoader, createUsersCollabContractsLoader, createCollabCollaboratorsLoader, createUsersOffersLoader } from "../DataLoaders/User"
+import { createUsersObjktLoader, createUsersLoader, createUsersCollabContractsLoader, createCollabCollaboratorsLoader, createUsersOffersLoader, createUsersGenerativeTokensLoader } from "../DataLoaders/User"
 
 
 
@@ -16,6 +16,7 @@ export interface RequestContext extends ExpressRequest {
 	userOffersLoader: ReturnType<typeof createUsersOffersLoader>,
 	userCollabContractsLoader: ReturnType<typeof createUsersCollabContractsLoader>,
 	collabCollaboratorsLoader: ReturnType<typeof createCollabCollaboratorsLoader>,
+  usersGenToksLoader: ReturnType<typeof createUsersGenerativeTokensLoader>,
 
   // GENERATIVE TOKEN loaders
   genTokLoader: ReturnType<typeof createGenTokLoader>,
